Asset and Level Update
Added more assets that can be used for the level. Also tweaked some
level design some more along with lighting just to get an idea of how
lighting functions in Unreal Engine 4 - going to have to play with that
more.

Have a To-Do list made up of a few more room tiles to make, as well as
UV mapping and texturing is on that list.

Also going to have to look into collision meshes as right now assets use
Complex Collision Meshes in order to get the map designed quicker.